requirements.txt
beautifulsoup4==4.10.0
PyYAML==6.0
requests==2.27.1
selectorlib==0.16.0
selenium==3.141.0
SQLAlchemy==1.4.31
